Asset-Management/WorkSphere/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
mysql-connector-python==9.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
qrcode[pil]==8.2
pydantic==2.11.4
pydantic-settings==2.9.1
email-validator==2.2.0
